requirements.txt
Django==4.2.7
python-decouple==3.8
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7